+++ engine/include/fury/fury.hpp
@@ -35,7 +35,7 @@ namespace fury {
 
 inline constexpr const char* engine_name() { return "Fury"; }
 inline constexpr int version_major() { return 2; }
-inline constexpr int version_minor() { return 1; }
+inline constexpr int version_minor() { return 2; }
 inline constexpr int version_patch() { return 0; }
 
 }  // namespace fury